network-automation_py/dhcp-script/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.2
bcrypt==3.2.0
Brlapi==0.6.6
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ciscoconfparse==1.5.19
colorama==0.4.3
command-not-found==0.3
cryptography==3.1
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dnspython==2.0.0
dogpile.cache==1.0.2
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
iso8601==0.1.13
isort==5.5.2
Jinja2==2.11.2
jmespath==0.10.0
jsonpatch==1.26
jsonpointer==2.0
junit-xml==1.9
junos-eznc==2.5.3
jxmlease==1.0.3
keyring==10.6.0
keyrings.alt==3.0
keystoneauth1==4.2.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
lxml==4.5.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
munch==2.5.0
napalm==3.2.0
ncclient==0.6.9
netaddr==0.8.0
netifaces==0.10.9
netmiko==3.3.0
ntc-templates==1.5.0
oauth==1.0.1
olefile==0.45.1
openstacksdk==0.49.0
os-client-config==2.1.0
os-service-types==1.7.0
paramiko==2.7.2
passlib==1.7.2
pathspec==0.8.0
pbr==5.5.0
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyeapi==0.8.3
pygobject==3.26.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-openstacksdk==0.5.2
pytz==2018.3
pyxdg==0.25
PyYAML==5.3.1
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
requestsexceptions==1.4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scp==0.13.2
SecretStorage==2.3.1
shade==1.33.0
simplejson==3.13.2
six==1.15.0
SQLAlchemy==1.3.19
stevedore==3.2.2
system-service==0.3
systemd-python==234
textfsm==1.1.0
toml==0.10.1
transitions==0.8.2
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
wadllib==1.3.2
wrapt==1.12.1
xkit==0.0.0
yamllint==1.24.2
yamlordereddictloader==0.4.0
zabbix-api==0.5.4
zipp==3.1.0
zope.interface==4.3.2
